# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Система управления учета оборудования и мебели",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/equipment-models": {
            "get": {
                "tags": [
                    "EquipmentModels"
                ],
                "summary": "Получить список всех моделей оборудования",
                "description": "Display a listing of the resource.",
                "operationId": "339cd0d57e62e92503212c0e34158432",
                "responses": {
                    "200": {
                        "description": "Список моделей оборудования",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/EquipmentModelIndexResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "EquipmentModels"
                ],
                "summary": "Создать новую модель оборудования",
                "description": "Store a newly created resource in storage.",
                "operationId": "882d49cb7214f692ef0e01535a5baa26",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEquipmentModelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Модель оборудования создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentModelShowResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/equipment-models/{id}": {
            "get": {
                "tags": [
                    "EquipmentModels"
                ],
                "summary": "Получить модель оборудования",
                "description": "Display the specified resource.",
                "operationId": "cd0953d27c0aca674c1c4bc607729423",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID модели оборудования",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Модель оборудования",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/EquipmentModelShowResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/equipment-models/{equipment_model}": {
            "delete": {
                "tags": [
                    "EquipmentModels"
                ],
                "summary": "Удалить модель оборудования",
                "description": "Remove the specified resource from storage.",
                "operationId": "037ab93c4fe6a9df4435d176d9be13c7",
                "parameters": [
                    {
                        "name": "equipment_model",
                        "in": "path",
                        "description": "ID модели оборудования",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Модель оборудования удалена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "EquipmentModels"
                ],
                "summary": "Обновить модель оборудования",
                "description": "Update the specified resource in storage.",
                "operationId": "100078a62dadc5aa9385306b44cb3651",
                "parameters": [
                    {
                        "name": "equipment_model",
                        "in": "path",
                        "description": "ID модели оборудования",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEquipmentModelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Модель оборудования изменена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentModelShowResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreEquipmentModelRequest": {
                "title": "Запрос на создание модели оборудования",
                "description": "Данные для создания новой модели оборудования",
                "required": [
                    "title",
                    "equipment_brand_id"
                ],
                "properties": {
                    "equipment_brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "EXP-123"
                    }
                },
                "type": "object"
            },
            "UpdateEquipmentModelRequest": {
                "title": "Запрос на обновление модели оборудования",
                "description": "Данные для обновления модели оборудования",
                "required": [
                    "title",
                    "equipment_brand_id"
                ],
                "properties": {
                    "equipment_brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "EXP-123"
                    }
                },
                "type": "object"
            },
...
            "EquipmentModelIndexResource": {
                "title": "Ресурс списка моделей оборудования",
                "description": "Формат данных моделей оборудования при получении списка",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "equipment_type": {
                                    "type": "string",
                                    "example": "HP"
                                },
                                "title": {
                                    "type": "string",
                                    "example": "EXP-123"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "EquipmentModelShowResource": {
                "title": "Ресурс модели оборудования",
                "description": "Формат данных модели оборудования",
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "equipment_brand": {
                                "type": "string",
                                "example": "HP"
                            },
                            "title": {
                                "type": "string",
                                "example": "EXP-123"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}